[AND-244] Encapsulate the promotions UI logic
Just move it to keep it together with the related logic code.
