+++ app-games/src/main/java/com/aptoide/android/aptoidegames/home/MainView.kt
@@ -26,14 +26,10 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import cm.aptoide.pt.extensions.animatedComposable
 import cm.aptoide.pt.extensions.staticComposable
-import cm.aptoide.pt.feature_campaigns.AptoideMMPCampaign
-import cm.aptoide.pt.feature_campaigns.toAptoideMMPCampaign
 import com.aptoide.android.aptoidegames.AptoideGamesBottomSheet
-import com.aptoide.android.aptoidegames.analytics.presentation.AnalyticsContext
 import com.aptoide.android.aptoidegames.apkfy.ApkfyBottomSheetContent
 import com.aptoide.android.aptoidegames.apkfy.presentation.rememberApkfyState
 import com.aptoide.android.aptoidegames.appview.appViewScreen
-import com.aptoide.android.aptoidegames.appview.buildAppViewRoute
 import com.aptoide.android.aptoidegames.appview.permissions.appPermissionsScreen
 import com.aptoide.android.aptoidegames.bottom_bar.AppGamesBottomBar
 import com.aptoide.android.aptoidegames.categories.presentation.allCategoriesScreen
@@ -49,9 +45,7 @@ import com.aptoide.android.aptoidegames.notifications.NotificationsPermissionReq
 import com.aptoide.android.aptoidegames.permissions.notifications.NotificationsPermissionViewModel
 import com.aptoide.android.aptoidegames.promo_codes.PromoCodeBottomSheet
 import com.aptoide.android.aptoidegames.promo_codes.rememberPromoCodeApp
-import com.aptoide.android.aptoidegames.promotions.analytics.rememberPromotionsAnalytics
 import com.aptoide.android.aptoidegames.promotions.presentation.PromotionDialog
-import com.aptoide.android.aptoidegames.promotions.presentation.PromotionsViewModel
 import com.aptoide.android.aptoidegames.search.presentation.searchScreen
 import com.aptoide.android.aptoidegames.settings.settingsScreen
 import com.aptoide.android.aptoidegames.theme.AptoideTheme
@@ -70,10 +64,6 @@ fun MainView(navController: NavHostController) {
   val coroutineScope = rememberCoroutineScope()
   val goBackHome: () -> Unit =
     { navController.popBackStack(navController.graph.startDestinationId, false) }
-
-  val promotionsViewModel = hiltViewModel<PromotionsViewModel>()
-  val promotionData by promotionsViewModel.uiState.collectAsState()
-  val promotionsAnalytics = rememberPromotionsAnalytics()
 
   val apkfyState = rememberApkfyState()
   var apkfyShown by remember { mutableStateOf(false) }
@@ -109,27 +99,7 @@ fun MainView(navController: NavHostController) {
             onDismiss = notificationsPermissionViewModel::dismissDialog
           )
         }
-        promotionData?.let {
-          AnalyticsContext.current.currentScreen = "home_dialog"
-          AptoideMMPCampaign.allowedBundleTags["home_dialog"] = "Ahab_v2" to it.first.uid
-          it.second.campaigns?.toAptoideMMPCampaign()?.sendImpressionEvent("home_dialog")
-          promotionsAnalytics.sendAhabV2DialogImpression(it.second.packageName)
-          PromotionDialog(
-            onPositiveClick = {
-              promotionsViewModel.dismissPromotion()
-              promotionsAnalytics.sendAhabV2DialogUpdate(it.second.packageName)
-              navController.navigateTo(buildAppViewRoute(it.second))
-            },
-            onNegativeClick = {
-              promotionsViewModel.dismissPromotion()
-              promotionsAnalytics.sendAhabV2DialogLater(it.second.packageName)
-            },
-            app = it.second,
-            title = it.first.title,
-            description = it.first.content,
-          )
-        }
-
+        PromotionDialog(navigate = navController::navigateTo)
 
         Box(modifier = Modifier.padding(padding)) {
           NavigationGraph(
